Add generate function for JD DGP

+++ DGPs/MA2.jl
@@ -5,7 +5,7 @@ end
 # ----- Model-specific utilities -----------------------------------------------
 insupport(::MA2, θ₁, θ₂) = (-2 < θ₁ < 2) && (-1 < θ₂ < 1) && (θ₂ - abs(θ₁) > -1)
 
-function simulate_ma2(θ₁::Float32, θ₂::Float32, n::Int64)
+function simulate_ma2(θ₁::Float32, θ₂::Float32, n::Int)
     ϵ = randn(Float32, n+2)
     ϵ[3:end] .+ θ₁ .* ϵ[2:end-1] .+ θ₂ .* ϵ[1:end-2]
 end
